Separation
Separated GlazeChemistry into ChemicalDatabase, MolecularDictionary, PeriodicTable, Element, Molecule, SimpleMolecule, and CompoundMolecule
Started on frits
