+++ GlazeBuilder/CompoundMolecules.json
@@ -1,12 +1,4 @@
 {
-	"Albite": {
-		"Formula": "[Na2O]:[Al2O3]:6[SiO2]",
-		"Submolecules": {
-			"Soda": 1,
-			"Alumina": 1,
-			"Silica": 6
-		}
-	},
 	"Alumina Hydrate": {
 		"Formula": "[Al2O3]:3[H2O]",
 		"Submolecules": {
@@ -23,6 +15,9 @@
 		}
 	},
 	"Bentonite": {
+		"Alternate Names": [
+			"Pyrophyllite"
+		],
 		"Formula": "[Al2O3]:4[SiO2]:[H2O]",
 		"Submolecules": {
 			"Alumina": 1,
@@ -31,6 +26,9 @@
 		}
 	},
 	"Bone Ash": {
+		"Alternate Names": [
+			"Calcium Phosphate"
+		],
 		"Formula": "Ca3:2[PO4]",
 		"Submolecules": {
 			"Phosphate": 2
@@ -74,6 +72,10 @@
 		}
 	},
 	"China Clay": {
+		"Alternate Names": [
+			"Raw Kaolonite Clay",
+			"Kaolin"
+		],
 		"Formula": "[Al2O3]:2[SiO2]:2[H2O]",
 		"Submolecules": {
 			"Alumina": 1,
@@ -105,6 +107,14 @@
 		},
 		"Copper": 1
 	},
+	"Cornwall Stone": {
+		"Formula": "[K2O]:[Al2O3]:8[SiO2]",
+		"Submolecules": {
+			"Potash": 1,
+			"Alumina": 1,
+			"Silica": 8
+		}
+	},
 	"Dolomite": {
 		"Formula": "[CaCO3]:[MgCO3]",
 		"Submolecules": {
@@ -112,7 +122,22 @@
 			"Magnesium Carbonate": 1
 		}
 	},
+	"Feldspar Lime": {
+		"Alternate Names": [
+			"Anorhite"
+		],
+		"Formula": "[CaO]:[Al2O3]:2[SiO2]",
+		"Submolecules": {
+			"Calcium Oxide": 1,
+			"Alumina": 1,
+			"Silica": 2
+		}
+	},
 	"Feldspar Potassium": {
+		"Alternate Names": [
+			"Feldspar Potash",
+			"Orthoclase"
+		],
 		"Formula": "[K2O]:[Al2O3]:6[SiO2]",
 		"Submolecules": {
 			"Potash": 1,
@@ -121,6 +146,10 @@
 		}
 	},
 	"Feldspar Sodium": {
+		"Alternate Names": [
+			"Albite",
+			"Feldspar Soda"
+		],
 		"Formula": "[Na2O]:[Al2O3]:6[SiO2]",
 		"Submolecules": {
 			"Soda": 1,
@@ -141,6 +170,13 @@
 		"Submolecules": {
 			"Iron Oxide Black": 1,
 			"Titanium Oxide": 1
+		}
+	},
+	"Kyanite": {
+		"Formula": "[Al2O3]:[SiO2]",
+		"Submolecules": {
+			"Alumina": 1,
+			"Silica": 1
 		}
 	},
 	"Lead Bisilicate": {
@@ -182,6 +218,13 @@
 		},
 		"Magnesium": 1
 	},
+	"Mullite": {
+		"Formula": "3[Al2O3]:2[SiO2]",
+		"Submolecules": {
+			"Alumina": 3,
+			"Silica": 2
+		}
+	},
 	"Nephiline Syenite": {
 		"Formula": "[K2O]:3[Na2O]:4[Al2O3]:8[SiO2]",
 		"Submolecules": {
@@ -191,28 +234,12 @@
 			"Silica": 8
 		}
 	},
-	"Orthoclase": {
-		"Formula": "[K2O]:[Al2O3]:6[SiO2]",
-		"Submolecules": {
-			"Potash": 1,
-			"Alumina": 1,
-			"Silica": 6
-		}
-	},
 	"Petalite": {
 		"Formula": "[Li2O]:[Al2O3]:8[SiO]",
 		"Submolecules": {
 			"Lithium Oxide": 1,
 			"Alumina": 1,
 			"Silica": 8
-		}
-	},
-	"Raw Kaolonite Clay": {
-		"Formula": "[Al2O3]:2[SiO2]:2[H2O]",
-		"Submolecules": {
-			"Alumina": 1,
-			"Silica": 2,
-			"Water": 2
 		}
 	},
 	"Sodalite": {
@@ -230,15 +257,11 @@
 			"Silica": 4
 		}
 	},
-	"Steatite": {
-		"Formula": "3[MgO]:4[SiO2]:[H2O]",
-		"Submolecules": {
-			"Magnesium Oxide": 3,
-			"Silica": 4,
-			"Water": 1
-		}
-	},
 	"Talc": {
+		"Alternate Names": [
+			"Magnesium Silicate",
+			"Steatite"
+		],
 		"Formula": "3[MgO]:4[SiO2]:[H2O]",
 		"Submolecules": {
 			"Magnesium Oxide": 3,
@@ -247,6 +270,9 @@
 		}
 	},
 	"Wollastonite": {
+		"Alternate Names": [
+			"Calcium Silicate"
+		],
 		"Formula": "[CaO]:[SiO2]",
 		"Submolecules": {
 			"Calcium Oxide": 1,
