+++ GlazeBuilder/MaterialsChemicals.json
@@ -1,0 +1,90 @@
+[
+	"Alumina",
+	"Alumina Hydrate",
+	"Amblygonite",
+	"Antimony Oxide",
+	"Barium Carbonate",
+	"Barium Oxide",
+	"Bentonite",
+	"Bone Ash",
+	"Borax",
+	"Boric Acid",
+	"Boric Oxide",
+	"Calcium Borate",
+	"Calcined Borax",
+	"Calcined Clay",
+	"Calcium Oxide",
+	"Cerium Oxide",
+	"China Clay",
+	"Chromium Oxide",
+	"Cobalt Carbonate",
+	"Cobalt Carbonate Basic",
+	"Cobalt Oxide",
+	"Colemanite",
+	"Copper Carbonate",
+	"Copper Carbonate Basic",
+	"Copper Oxide Black",
+	"Copper Oxide Red",
+	"Cornwall Stone",
+	"Crocus Martis",
+	"Cryolite",
+	"Dolomite",
+	"Erbium Oxide",
+	"Feldspar Lime",
+	"Feldspar Potassium",
+	"Feldspar Sodium",
+	"Fluorspar",
+	"Gerstley Borate",
+	"Ilmenite",
+	"Iron Chromate",
+	"Iron Oxide Black",
+	"Iron Oxide Red",
+	"Iron Oxide",
+	"Kyanite",
+	"Lead Bisilicate",
+	"Lead Carbonate White",
+	"Lead Oxide Red",
+	"Lead Oxide Yellow",
+	"Lead Sesquisilicate",
+	"Lead Sulphide",
+	"Lithium Carbonate",
+	"Lithium Oxide",
+	"Magnesium Carbonate",
+	"Magnesium Oxide",
+	"Manganese Carbonate",
+	"Manganese Carbonate Light",
+	"Manganese Dioxide",
+	"Mullite",
+	"Neodymium Oxide",
+	"Nepheline Syenite"
+	"Nickel Oxide Green",
+	"Nickel Oxide Black",
+	"Petalite",
+	"Potash",
+	"Potassium Carbonate",
+	"Potassium Chromate",
+	"Potassium Dichromate",
+	"Potassium Nitrate",
+	"Praseodymium Oxide",
+	"Silica",
+	"Silicon Carbide",
+	"Soda",
+	"Sodalite",
+	"Sodium Bicarbonate",
+	"Sodium Carbonate",
+	"Spinel",
+	"Spodumene",
+	"Strontium Carbonate",
+	"Strontium Oxide",
+	"Talc",
+	"Tin Oxide White",
+	"Titanium Oxide",
+	"Uranium Oxide",
+	"Vanadium Oxide",
+	"Whiting",
+	"Wollastonite",
+	"Yellow Ocher",
+	"Zinc Oxide",
+	"Zirconium Oxide",
+	"Zirconium Silicate"
+]
