+++ GlazeBuilder/MaterialsFrits.json
@@ -1,0 +1,25 @@
+{
+	"Ferro Frit 3110-2": {
+		"Fluxes": {
+			"Potash": ,
+			"Soda": ,
+			"Calcium Oxide": ,
+			"Magnesium Oxide": ,
+			"Barium Oxide": ,
+			"Zinc Oxide": ,
+			"Lithium Oxide": ,
+			"Strontium Oxide": ,
+			"Lead Oxide": 
+		},
+		"Non-Fluxes:": {
+			"Alumina": ,
+			"Boric Oxide": ,
+			"Zirconium Dioxide": ,
+			"Fluorine": ,
+			"Silica": 
+		}
+	},
+	"Frit 3124-2": {
+
+	}
+}
